Action to inject instance of type
To inject a service, write it's name, right click and select `Inject instance`.

Tried implementing as a refactoring provider first, but couldn't get that work when running as a NuGet package. This will instead register a fixer for when you try to use a class or interface as a variable.